requirements.txt
chromadb~=0.4.24
openai~=1.16.2
pydantic~=2.6.4
python-dotenv~=1.0.1
streamlit~=1.33.0
requests~=2.31.0
duckduckgo_search~=5.2.2
pillow~=10.3.0
transformers~=4.39.3
